requirements.txt
FPDF
Pillow
PyYAML
